% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_file_size.R
\name{get_all_file_size}
\alias{get_all_file_size}
\title{get_all_file_size Title
Elle prend un chemin de dossier en paramètre, et map la fonction get_one_file_size pour n’obtenir qu’un seul tableau.}
\usage{
get_all_file_size(directory)
}
\arguments{
\item{directory}{un chemin de dossier}
}
\value{
1 tableau
}
\description{
get_all_file_size Title
Elle prend un chemin de dossier en paramètre, et map la fonction get_one_file_size pour n’obtenir qu’un seul tableau.
}
\examples{
library(purrr)
library(dplyr)
get_all_file_size(directory = ".")
}
